requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
jwt==1.2.0
pycparser==2.20
pytz==2021.1
quickemailverification==1.0.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
